Moved mapbox class to correct package, change bottom menu map link
