+++ app/src/main/res/layout/activity_mapbox.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="br.com.ecostage.mobilecollect.MapboxActivity">
+    tools:context="br.com.ecostage.mobilecollect.ui.map.MapboxActivity">
 
     <android.support.design.widget.CoordinatorLayout
         android:id="@+id/mapContainerView"
